refactor: move proof verification to crypto crate
